requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.32,<1.0
google-genai>=1.0,<2.0
python-dotenv>=1.0,<2.0
